+++ rest_server/database/__init__.py
@@ -1,0 +1,3 @@
+"""Init."""
+
+from . import interface, schema  # noqa: F401
